refactor(fbuild-build): split compile_database.rs into submodules to satisfy LOC gate
